Adding DMARC TXT
Without DMARC, it is possible for anyone to craft an email from our mail server. The best option would be to set the `p`=`reject`  but, as explained [in this link](https://support.google.com/a/answer/2466563?hl=es), it may be kind of risky to do this without first analyzing if there is some kind of extra configuration we need to do for allowing certain exceptions.